change units for municipality emissions to co2 and organize components
