+++ src/components/municipalities/emissionsGraph/SectorsChart.tsx
