+++ src/locales/en/translation.json
@@ -710,7 +710,7 @@
     "annualChangeSince2015": "Annual emission change since 2015",
     "reductionToMeetParis": "Emission reduction to meet the Paris Agreement",
     "cannotReduceToParis": "Will not reach Paris Agreement",
-    "consumptionEmissionsPerCapita": "Consumption emissions per capita in tons CO₂",
+    "consumptionEmissionsPerCapita": "Consumption emissions per capita in tons CO₂e",
     "climatePlan": "Climate plan",
     "noPlan": "No plan",
     "noClimatePlan": "No climate plan",
