+++ src/locales/sv/translation.json
@@ -714,7 +714,7 @@
     "annualChangeSince2015": "Årlig utsläppsförändring sedan 2015",
     "reductionToMeetParis": "Utsläppsminskning för att klara Parisavtalet",
     "cannotReduceToParis": "Klarar ej Parisavtalet",
-    "consumptionEmissionsPerCapita": "Konsumtionsutsläpp per invånare i ton CO₂",
+    "consumptionEmissionsPerCapita": "Konsumtionsutsläpp per invånare i ton CO₂e",
     "climatePlan": "Klimatplan",
     "noPlan": "Saknar plan",
     "noClimatePlan": "Saknar klimatplan",
